refactor: [step4] Tail/Body Strategy를 하나로 합침
