+++ src/main/java/ladder/step4/domain/strategy/DirectionStrategy.java
@@ -3,5 +3,6 @@ package ladder.step4.domain.strategy;
 import ladder.step4.domain.Direction;
 
 public interface DirectionStrategy {
-  Direction create(Direction prev);
+  Direction createBody(Direction prev);
+  Direction createTail(Direction prev);
 }
